Move Demo File function
